+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod apply;
+pub mod get_status;
 pub mod get_targets;
 pub mod parse;
 pub mod test_move;